backend/requirements.txt
alabaster==0.7.12
asgiref==3.4.1
Babel==2.9.1
certifi==2021.10.8
charset-normalizer==2.0.10
commonmark==0.9.1
dj-database-url==0.5.0
Django==3.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
docutils==0.17.1
idna==3.3
imagesize==1.3.0
Jinja2==3.0.3
MarkupSafe==2.0.1
packaging==21.3
psycopg2==2.9.3
Pygments==2.11.2
PyJWT==2.3.0
pyparsing==3.0.6
python-decouple==3.5
pytz==2021.3
recommonmark==0.7.1
requests==2.27.1
snowballstemmer==2.2.0
Sphinx==4.3.2
sphinx-rtd-theme==1.0.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.2
urllib3==1.26.8
